Added JWT Authentication from scratch

+++ src/main/java/com/major/project/airBnbApp/service/HotelServiceImpl.java
@@ -6,12 +6,15 @@ import com.major.project.airBnbApp.dto.HotelInfoDto;
 import com.major.project.airBnbApp.dto.RoomDto;
 import com.major.project.airBnbApp.entity.Hotel;
 import com.major.project.airBnbApp.entity.Room;
+import com.major.project.airBnbApp.entity.User;
 import com.major.project.airBnbApp.exception.ResourceNotFoundException;
+import com.major.project.airBnbApp.exception.UnauthorizedException;
 import com.major.project.airBnbApp.repository.HotelRepository;
 import com.major.project.airBnbApp.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -29,28 +32,40 @@ public class HotelServiceImpl implements HotelService{
     private final RoomRepository roomRepository;
 
     @Override
+    @Transactional
     public HotelDto createNewHotel(HotelDto hotelDto) {
         log.info("Creating a new hotel with name: {}", hotelDto.getName());
         Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
         hotel.setActive(false);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
         hotel = hotelRepository.save(hotel);
         log.info("Created a new hotel with ID: {}", hotel.getId());
         return modelMapper.map(hotel, HotelDto.class);
     }
 
     @Override
-    public HotelDto getHotelById(Long id) {
+    public HotelDto getHotelById(Long id) throws UnauthorizedException {
         log.info("Getting the Hotel with Id: {}", id);
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with Id: "+ id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnauthorizedException("This user does not own this hotel with id: "+ id);
+        }
         return modelMapper.map(hotel, HotelDto.class);
     }
 
     @Override
-    public HotelDto updateHotelById(Long id, HotelDto hotelDto) {
+    @Transactional
+    public HotelDto updateHotelById(Long id, HotelDto hotelDto) throws UnauthorizedException {
         log.info("Getting the hotel for update.");
         Hotel hotel = hotelRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Hotel not found with this id :"+id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnauthorizedException("This user does not own this hotel with id: "+ id);
+        }
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
         hotel = hotelRepository.save(hotel);
@@ -60,10 +75,14 @@ public class HotelServiceImpl implements HotelService{
 
     @Transactional
     @Override
-    public void DeleteHotelById(Long id) {
+    public void DeleteHotelById(Long id) throws UnauthorizedException {
         Hotel hotel = hotelRepository.findById(id).orElseThrow(
                 () -> new ResourceNotFoundException("Hotel not found with id: "+ id));
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnauthorizedException("This user does not own this hotel with id: "+ id);
+        }
         for(Room room: hotel.getRooms()){
             inventoryService.deleteAllInventories(room);
             roomRepository.deleteById(room.getId());
@@ -73,10 +92,14 @@ public class HotelServiceImpl implements HotelService{
 
     @Override
     @Transactional
-    public void activateHotel(Long hotelId) {
+    public void activateHotel(Long hotelId) throws UnauthorizedException {
         log.info("Activating the hotel...with Id: {}", hotelId);
         Hotel hotel = hotelRepository.findById(hotelId).orElseThrow(
                 () -> new ResourceNotFoundException("Hotel not found with id: "+ hotelId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnauthorizedException("This user does not own this hotel with id: "+ hotelId);
+        }
         hotel.setActive(true);
         // Assuming only do it once
         for(Room room: hotel.getRooms()){
@@ -85,7 +108,7 @@ public class HotelServiceImpl implements HotelService{
 //        modelMapper.map(hotel, HotelDto)
     }
 
-
+//public methods
     @Override
     public List<HotelDto> getAllHotels() {
         log.info("Getting All Hotels");
