+++ src/main/java/com/major/project/airBnbApp/service/RoomServiceImpl.java
@@ -3,12 +3,15 @@ package com.major.project.airBnbApp.service;
 import com.major.project.airBnbApp.dto.RoomDto;
 import com.major.project.airBnbApp.entity.Hotel;
 import com.major.project.airBnbApp.entity.Room;
+import com.major.project.airBnbApp.entity.User;
 import com.major.project.airBnbApp.exception.ResourceNotFoundException;
+import com.major.project.airBnbApp.exception.UnauthorizedException;
 import com.major.project.airBnbApp.repository.HotelRepository;
 import com.major.project.airBnbApp.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -27,9 +30,14 @@ public class RoomServiceImpl implements RoomService{
 
 
     @Override
-    public RoomDto createNewRoom(Long hotelId, RoomDto roomDto) {
+    @Transactional
+    public RoomDto createNewRoom(Long hotelId, RoomDto roomDto) throws UnauthorizedException {
         log.info("Creating a new room in hotel with ID: {}", hotelId);
         Hotel hotel = hotelRepository.findById(hotelId).orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id: "+ hotelId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnauthorizedException("This user does not own this hotel with id: "+ hotelId);
+        }
         Room room = modelMapper.map(roomDto, Room.class);
         room.setHotel(hotel);
         room = roomRepository.save(room);
@@ -42,16 +50,21 @@ public class RoomServiceImpl implements RoomService{
 
     @Transactional
     @Override
-    public List<RoomDto> getAllRoomsInHotel(Long hotelId) {
+    public List<RoomDto> getAllRoomsInHotel(Long hotelId) throws UnauthorizedException {
         log.info("Getting All Rooms from "+ hotelId + " Hotel");
         Hotel hotel = hotelRepository.findById(hotelId)
                 .orElseThrow(
                         () -> new ResourceNotFoundException("Hotel not found with id: "+ hotelId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnauthorizedException("This user does not own this hotel with id: "+ hotelId);
+        }
         return hotel.getRooms()
                 .stream()
                 .map((e) -> modelMapper.map(e, RoomDto.class))
                 .collect(Collectors.toList());
     }
+
 
     @Override
     public RoomDto getRoomById(Long roomId) {
@@ -64,11 +77,15 @@ public class RoomServiceImpl implements RoomService{
 
     @Transactional
     @Override
-    public void deleteRoomById(Long roomId) {
+    public void deleteRoomById(Long roomId) throws UnauthorizedException {
         log.info("Deleting the room with Id: " + roomId);
         boolean exists = roomRepository.existsById(roomId);
         Room room = roomRepository.findById(roomId).orElseThrow(
                 () -> new ResourceNotFoundException("Hotel not found with id: "+ roomId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner())){
+            throw new UnauthorizedException("This user does not own this hotel with id: "+ roomId);
+        }
         inventoryService.deleteAllInventories(room);
         roomRepository.deleteById(roomId);
     }
